Improve the lower case x.
This is the new preferred way to handle diagonals.
These segments are all at 45°.
These points line up perfectly on the grid.
The normal version of the line is slightly thinner
than horizontal or vertical lines.
The bold version of the line is slightly thicker.
Neither is obvious at a normal size.

I don't know why ^ keeps changing!
I haven't touched it.
It will be replaced soon.